update: - extracted the session filters to utils - cleaned up some code in the filter modal
