+++ src/utils/tapIndex.js
@@ -4,3 +4,4 @@ export * from './models'
 export * from './dateTime'
 export * from './theme'
 export * from './logs'
+export * from './filters'
